refactor: Card 컴포넌트로 API 데이터 받아오기 구현
