+++ src/components/Card/Card.jsx
@@ -1,30 +1,22 @@
 import React, { useState, useEffect } from 'react';
 import Moment from 'moment';
-import axios from 'axios';
+
 import { getElapsedTime } from '../../utils/getElapsedTime';
 import './Card.css';
 
-export default function Card() {
+export default function Card({ data }) {
   const [cardData, setCardData] = useState([]);
 
   useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await axios.get('https://bootcamp-api.codeit.kr/api/sample/folder');
-        setCardData(response.data.folder.links);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-    };
+    setCardData(data);
+  }, [data]);
 
-    fetchData();
-  }, []);
   return (
     <>
       {cardData === null ? (
         <p id="mainDefaultText">저장된 링크가 없습니다.</p>
       ) : (
-        cardData.map(link => {
+        cardData.map(link => (
           <div key={link.id} className="card">
             <a href={link.url}>
               <img
@@ -38,8 +30,8 @@ export default function Card() {
                 <div className="uploadDate">{Moment(link.createdAt).format('YYYY.MM.DD')}</div>
               </div>
             </a>
-          </div>;
-        })
+          </div>
+        ))
       )}
     </>
   );
