+++ src/components/pages/shared.jsx
@@ -1,9 +1,8 @@
 import Header from '../Header/Header';
 import Section from '../Section/Section';
-import Main from '../CardList/CardList';
+import CardList from '../CardList/CardList';
 import Footer from '../Footer/Footer';
 import SearchBar from '../SearchBar/SearchBar';
-import CardList from '../CardList/CardList';
 
 export default function Shared() {
   return (
